Abstracts out datetime modal picker
